Initial setup. Retrieves recipes and shows them in the RecyclerView.

+++ app/src/main/java/it/redlor/bakingapp/pojos/Ingredient.java
@@ -11,7 +11,7 @@ public class Ingredient {
 
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private float quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -19,11 +19,11 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public Integer getQuantity() {
+    public float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
